Patch over buggy library behavior in Julia CI
A recent fix for H5Oexists(_by_name) changes the returned value of
the API call when checking for a non-existent object with a file
ID (FAIL should be false). The Julia CI explicitly tests for the
buggy result.

We've notified the maintainers and this source patch will fix
the issue for our CI until it is fixed upstream.